Move Shift logging at the data layer
