+++ app/src/androidTest/java/com/github/factotum_sdp/factotum/utils/DeliveryLoggerUtils.kt
@@ -1,7 +1,8 @@
 package com.github.factotum_sdp.factotum.utils
 
 import com.github.factotum_sdp.factotum.data.DeliveryLogger
-import com.github.factotum_sdp.factotum.data.FirebaseInstance
+import com.github.factotum_sdp.factotum.firebase.FirebaseInstance
+import com.github.factotum_sdp.factotum.firebase.FirebaseStringFormat
 import com.github.factotum_sdp.factotum.ui.roadbook.DRecordList
 
 class DeliveryLoggerUtils {
