+++ app/src/androidTest/java/com/github/factotum_sdp/factotum/utils/FirebaseStringFormatTest.kt
@@ -1,6 +1,7 @@
 package com.github.factotum_sdp.factotum.utils
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.factotum_sdp.factotum.firebase.FirebaseStringFormat
 import org.junit.Test
 import org.junit.runner.RunWith
