+++ app/src/main/java/com/github/factotum_sdp/factotum/data/LoginDataSource.kt
@@ -1,5 +1,6 @@
 package com.github.factotum_sdp.factotum.data
 
+import com.github.factotum_sdp.factotum.firebase.FirebaseInstance
 import com.github.factotum_sdp.factotum.models.Role
 import com.github.factotum_sdp.factotum.models.User
 import com.google.firebase.auth.UserProfileChangeRequest
