+++ app/src/main/java/com/github/factotum_sdp/factotum/data/SignUpDataSink.kt
@@ -1,6 +1,7 @@
 package com.github.factotum_sdp.factotum.data
 
 import com.github.factotum_sdp.factotum.data.LoginDataSource.Companion.DISPATCH_DB_PATH
+import com.github.factotum_sdp.factotum.firebase.FirebaseInstance
 import com.github.factotum_sdp.factotum.models.User
 import java.io.IOException
 import java.util.concurrent.CompletableFuture
