+++ app/src/main/java/com/github/factotum_sdp/factotum/ui/signup/SignUpFragment.kt
@@ -17,7 +17,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.github.factotum_sdp.factotum.R
 import com.github.factotum_sdp.factotum.data.*
-import com.github.factotum_sdp.factotum.data.FirebaseInstance.getAuth
+import com.github.factotum_sdp.factotum.firebase.FirebaseInstance.getAuth
 import com.github.factotum_sdp.factotum.databinding.FragmentSignupBinding
 import com.github.factotum_sdp.factotum.models.Role
 import com.github.factotum_sdp.factotum.models.User
